created new category controller & moved it to admin directory with autorize filter and route /admin
